1. 전투력 Double형으로 변경 2. 메일 전송 추가 3. 개발서버 설정

+++ IdleGameServer/src/main/java/com/onlyonegames/eternalfantasia/domain/model/dto/Contents/PreviousBattlePowerRankingDto.java
@@ -10,7 +10,7 @@ public class PreviousBattlePowerRankingDto {
     Long id;
     Long useridUser;
     String userGameName;
-    Long battlePower;
+    double battlePower;
     int ranking;
     boolean receivable;
